fix: impossible to use card with calendar
